removed useless and fixed rest

+++ ft_memchr.c
@@ -6,7 +6,7 @@
 /*   By: guiricha <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/11/24 18:36:35 by guiricha          #+#    #+#             */
-/*   Updated: 2015/11/27 20:50:24 by guiricha         ###   ########.fr       */
+/*   Updated: 2015/12/09 14:20:44 by guiricha         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void				*ft_memchr(const void *s, int c, size_t n)
 {
-	unsigned int	i;
+	size_t			i;
 	char			*str;
 
 	str = (char *)s;
